RestApisV2/requirements.txt
Django~=4.0.2
djangorestframework~=3.13.1
django-ckeditor~=6.2.0
PyMySQL~=1.0.2